# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Prints a URLs Dashboard",
                "description": "Prints a URLs Dashboard",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful dashboard print",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string",
                                    "example": "<h1>URL Shortener</h1>"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string",
                                    "example": "<h1>Bad Request</h1>"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Internal Error",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string",
                                    "example": "<h1>Server Internal Error</h1>"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/generate-short-url": {
            "post": {
                "tags": [
                    "URL Shortener"
                ],
                "summary": "Generate a short URL",
                "description": "Generate a short URL",
                "operationId": "generateShortUrl",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "example": "https://brayanibp.dev"
                                    },
                                    "db_shard": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "short_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/emPjo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Internal Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Internal Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/redirect/{shortenedUrl}": {
            "get": {
                "tags": [
                    "URL Shortener"
                ],
                "summary": "Show redirect page",
                "description": "Show redirect page",
                "operationId": "showRedirect",
                "parameters": [
                    {
                        "name": "shortenedUrl",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "emPjo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string",
                                    "example": "<h1>Redirect</h1>"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string",
                                    "example": "<h1>Bad Request</h1>"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Internal Error",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string",
                                    "example": "<h1>Server Internal Error</h1>"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/{short_url}": {
            "get": {
                "tags": [
                    "URL Shortener"
                ],
                "summary": "Redirect to original URL",
                "description": "Redirect to original URL",
                "operationId": "redirectToOriginalUrl",
                "parameters": [
                    {
                        "name": "short_url",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "emPjo"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "example": "https://brayanibp.dev"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "URL not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Home",
            "description": "Home"
        },
        {
            "name": "URL Shortener",
            "description": "URL Shortener"
        }
    ]
}